refactor(provider): own provider-domain logic in pkg/provider
Provider-source resolution, downloading, renaming, and version
listing/caching previously lived in pkg/workspace alongside actual
workspace orchestration. Move them into pkg/provider so workspace
keeps only the operations that depend on workspace internals
(LoadProviders, AddProvider, UpdateProvider, SwitchProvider,
ListProviderVersions/SetProviderVersion/CheckAllProviderVersions,
CheckProviderUpdate). Also drop the redundant provider_ prefix
from filenames inside pkg/provider.
