+++ cmd/helper/get_provider_name.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/devsy-org/devsy/cmd/flags"
 	"github.com/devsy-org/devsy/pkg/provider"
-	"github.com/devsy-org/devsy/pkg/workspace"
 	"github.com/spf13/cobra"
 )
 
@@ -36,7 +35,7 @@ func (cmd *GetProviderNameCmd) Run(ctx context.Context, args []string) error {
 		return fmt.Errorf("provider is missing")
 	}
 
-	providerRaw, _, err := workspace.ResolveProvider(args[0])
+	providerRaw, _, err := provider.ResolveProvider(args[0])
 	if err != nil {
 		return fmt.Errorf("resolve provider: %w", err)
 	}
