+++ pkg/provider/resolve_test.go
@@ -1,9 +1,7 @@
-package workspace
+package provider
 
 import (
 	"testing"
-
-	"github.com/devsy-org/devsy/pkg/provider"
 )
 
 func TestGetProviderSource_GithubCanonicalization(t *testing.T) {
@@ -12,41 +10,41 @@ func TestGetProviderSource_GithubCanonicalization(t *testing.T) {
 
 	tests := []struct {
 		name string
-		src  provider.ProviderSource
+		src  ProviderSource
 		want string
 	}{
 		{
 			name: "bare org/repo gets github.com prefix",
-			src:  provider.ProviderSource{Github: githubPath},
+			src:  ProviderSource{Github: githubPath},
 			want: canonicalGithub,
 		},
 		{
 			name: "already-prefixed github source is preserved",
-			src:  provider.ProviderSource{Github: canonicalGithub},
+			src:  ProviderSource{Github: canonicalGithub},
 			want: canonicalGithub,
 		},
 		{
 			name: "internal source returns raw",
-			src:  provider.ProviderSource{Internal: true, Raw: "docker"},
+			src:  ProviderSource{Internal: true, Raw: "docker"},
 			want: "docker",
 		},
 		{
 			name: "url source preserved",
-			src:  provider.ProviderSource{URL: "https://example.com/p.yaml"},
+			src:  ProviderSource{URL: "https://example.com/p.yaml"},
 			want: "https://example.com/p.yaml",
 		},
 		{
 			name: "file source preserved",
-			src:  provider.ProviderSource{File: "/tmp/p.yaml"},
+			src:  ProviderSource{File: "/tmp/p.yaml"},
 			want: "/tmp/p.yaml",
 		},
 	}
 
 	for _, tc := range tests {
 		t.Run(tc.name, func(t *testing.T) {
-			got := getProviderSource(tc.src, "config-name")
+			got := GetProviderSource(tc.src, "config-name")
 			if got != tc.want {
-				t.Fatalf("getProviderSource() = %q, want %q", got, tc.want)
+				t.Fatalf("GetProviderSource() = %q, want %q", got, tc.want)
 			}
 		})
 	}
