+++ pkg/provider/versions_testconst_test.go
@@ -1,0 +1,7 @@
+package provider
+
+const (
+	testTagV100 = "v1.0.0"
+	testNameABC = "abc"
+	testNameBar = "bar"
+)
