+++ pkg/workspace/provider_versions_test.go
@@ -1,123 +1,38 @@
 package workspace
 
 import (
-	"encoding/json"
 	"errors"
 	"testing"
 	"time"
+
+	"github.com/devsy-org/devsy/pkg/provider"
 )
-
-func TestErrVersionListUnsupported(t *testing.T) {
-	if ErrVersionListUnsupported == nil {
-		t.Fatal("ErrVersionListUnsupported must be defined")
-	}
-	wrapped := errors.New("wrapped: " + ErrVersionListUnsupported.Error())
-	if !errors.Is(errors.Join(ErrVersionListUnsupported, wrapped), ErrVersionListUnsupported) {
-		t.Fatal("errors.Is must work against ErrVersionListUnsupported")
-	}
-}
-
-func TestProviderVersionFields(t *testing.T) {
-	v := ProviderVersion{Tag: testTagV100, Current: true}
-	if v.Tag != testTagV100 || !v.Current {
-		t.Fatal("fields must round-trip")
-	}
-}
-
-func TestClassifyVersionSource(t *testing.T) {
-	cases := []struct {
-		in   string
-		kind sourceKind
-	}{
-		{"github.com/devsy-org/devsy-provider-aws@v1.2.0", sourceGitHub},
-		{"github.com/devsy-org/devsy-provider-aws", sourceGitHub},
-		{"https://example.com/foo/provider.yaml", sourceManifestURL},
-		{"https://example.com/foo/provider.yaml@v1.0.0", sourceManifestURL},
-		{"/abs/path/provider.yaml", sourceLocal},
-		{"./relative/provider.yaml", sourceLocal},
-	}
-	for _, c := range cases {
-		t.Run(c.in, func(t *testing.T) {
-			got := classifyVersionSource(c.in)
-			if got != c.kind {
-				t.Fatalf("got %v, want %v", got, c.kind)
-			}
-		})
-	}
-}
-
-func TestListVersionsForSource_LocalUnsupported(t *testing.T) {
-	_, err := listVersionsForSource("/abs/path/provider.yaml", ListVersionsOptions{})
-	if !errors.Is(err, ErrVersionListUnsupported) {
-		t.Fatalf("local source must be unsupported, got %v", err)
-	}
-}
-
-func TestListVersionsForSource_UnknownUnsupported(t *testing.T) {
-	_, err := listVersionsForSource("totally-bogus-source", ListVersionsOptions{})
-	if !errors.Is(err, ErrVersionListUnsupported) {
-		t.Fatalf("unknown source must be unsupported, got %v", err)
-	}
-}
-
-func TestListVersionsForSource_GitHubInvalid(t *testing.T) {
-	_, err := listVersionsForSource("github.com/missingrepo", ListVersionsOptions{})
-	if err == nil || errors.Is(err, ErrVersionListUnsupported) {
-		t.Fatalf("github source missing repo segment must error (not unsupported): %v", err)
-	}
-}
-
-func TestMarkCurrent(t *testing.T) {
-	versions := []ProviderVersion{{Tag: testTagV100}, {Tag: "v0.9.0"}}
-	got := markCurrent(versions, "github.com/foo/bar@v1.0.0")
-	if !got[0].Current || got[1].Current {
-		t.Fatalf("only v1.0.0 should be marked current: %+v", got)
-	}
-}
-
-func TestMarkCurrent_NoTag(t *testing.T) {
-	versions := []ProviderVersion{{Tag: testTagV100}}
-	got := markCurrent(versions, "github.com/foo/bar")
-	if got[0].Current {
-		t.Fatal("no pinned tag → none current")
-	}
-}
-
-func TestRewriteSourceTag(t *testing.T) {
-	got, err := rewriteSourceTag("github.com/foo/bar@v1.0.0", "v2.0.0")
-	if err != nil || got != "github.com/foo/bar@v2.0.0" {
-		t.Fatalf("got %q err %v", got, err)
-	}
-	got, err = rewriteSourceTag("github.com/foo/bar", "v2.0.0")
-	if err != nil || got != "github.com/foo/bar@v2.0.0" {
-		t.Fatalf("got %q err %v", got, err)
-	}
-	if _, err := rewriteSourceTag("github.com/foo/bar", ""); err == nil {
-		t.Fatal("empty tag must error")
-	}
-}
 
 func TestListVersionsForSource_CachesResults(t *testing.T) {
 	dir := t.TempDir()
 	t.Setenv("DEVSY_HOME", dir)
 
 	source := "github.com/foo/bar@v1.0.0"
-	hash := hashProviderSource(source)
+	hash := provider.HashProviderSource(source)
 
 	// Prime cache with a synthetic entry that doesn't match any real upstream.
-	cached := providerVersionCache{
+	cached := provider.ProviderVersionCache{
 		"myprov": {
 			SourceHash: hash,
-			Versions:   []ProviderVersion{{Tag: testTagV999}},
+			Versions:   []provider.ProviderVersion{{Tag: testTagV999}},
 			FetchedAt:  time.Now(),
 		},
 	}
-	if err := SaveProviderVersionCache(cached); err != nil {
+	if err := provider.SaveProviderVersionCache(cached); err != nil {
 		t.Fatal(err)
 	}
 
 	// listVersionsForSourceCached must read the cache when UseCache is set and the entry is fresh.
-	got, err := listVersionsForSourceCached("myprov", source, ListVersionsOptions{UseCache: true})
+	got, err := listVersionsForSourceCached(
+		"myprov",
+		source,
+		provider.ListVersionsOptions{UseCache: true},
+	)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -131,53 +46,29 @@ func TestListVersionsForSource_BypassesCache(t *testing.T) {
 	t.Setenv("DEVSY_HOME", dir)
 
 	source := "/local/path/provider.yaml"
-	hash := hashProviderSource(source)
-	cached := providerVersionCache{
+	hash := provider.HashProviderSource(source)
+	cached := provider.ProviderVersionCache{
 		"myprov": {
 			SourceHash: hash,
-			Versions:   []ProviderVersion{{Tag: testTagV999}},
+			Versions:   []provider.ProviderVersion{{Tag: testTagV999}},
 			FetchedAt:  time.Now(),
 		},
 	}
-	if err := SaveProviderVersionCache(cached); err != nil {
+	if err := provider.SaveProviderVersionCache(cached); err != nil {
 		t.Fatal(err)
 	}
 
 	// With UseCache=false the cache is ignored and the underlying classifier runs.
 	// Local source → ErrVersionListUnsupported.
-	_, err := listVersionsForSourceCached("myprov", source, ListVersionsOptions{UseCache: false})
-	if !errors.Is(err, ErrVersionListUnsupported) {
+	_, err := listVersionsForSourceCached(
+		"myprov",
+		source,
+		provider.ListVersionsOptions{UseCache: false},
+	)
+	if !errors.Is(err, provider.ErrVersionListUnsupported) {
 		t.Fatalf(
 			"expected ErrVersionListUnsupported when bypassing cache for local source, got %v",
 			err,
 		)
-	}
-}
-
-func TestProviderVersionCheckResult_UnsupportedShape(t *testing.T) {
-	// Verify the struct shape and JSON tags by marshalling.
-	r := ProviderVersionCheckResult{
-		Current:     testTagV100,
-		Unsupported: true,
-	}
-	data, err := json.Marshal(r)
-	if err != nil {
-		t.Fatal(err)
-	}
-	expected := `{"current":"v1.0.0","latest":"","updateAvailable":false,"unsupported":true}`
-	if string(data) != expected {
-		t.Fatalf("JSON shape wrong:\ngot:  %s\nwant: %s", data, expected)
-	}
-}
-
-func TestProviderVersionCheckResult_ErrorShape(t *testing.T) {
-	r := ProviderVersionCheckResult{Error: "boom"}
-	data, err := json.Marshal(r)
-	if err != nil {
-		t.Fatal(err)
-	}
-	expected := `{"current":"","latest":"","updateAvailable":false,"unsupported":false,"error":"boom"}`
-	if string(data) != expected {
-		t.Fatalf("JSON shape wrong:\ngot:  %s\nwant: %s", data, expected)
 	}
 }
