+++ pkg/workspace/provider_versions_testconst_test.go
@@ -1,8 +1,5 @@
 package workspace
 
 const (
-	testTagV100 = "v1.0.0"
 	testTagV999 = "v9.9.9"
-	testNameABC = "abc"
-	testNameBar = "bar"
 )
